use project lifecycle events for operations. Register to postprovision
